Nutrition to Double

+++ MicroNutritionist/MicroNutritionist/Common/Database/Models/ProductNutritionAmount.cs
@@ -16,6 +16,6 @@ namespace MicroNutritionist.Common.Database.Models
         public int ProductId { get; set; }
         [Indexed(Unique = false)]
         public int NutritionId { get; set; }
-        public int AmountMg { get; set; }
+        public double AmountMg { get; set; }
     }
 }
